Smart_Agriculture_System-main/smart_agriculture_api/requirements.txt
Flask
flask-cors
tensorflow
numpy
Pillow
requests
joblib
scikit-learn